Companies Ranked! (2021)
Companies ranked based on their attributes

This file contains data on 846 companies and their ranking based on 6 characteristics: customer satisfaction, employee engagement and development, innovation, social responsibility, financial strength, and effectiveness. These rank were made by the Drucker Institute. Obviously these rankings are not and can not be accurate, but are the opinion of and are influenced wholly by the ranking criterium of the Drucker Institute.